Previously, a range was a type with any of: * `begin` + size` * `begin` + `end` * `before_begin` + `end`
Now it can also be `data` + `size`.

This change means that `containers::data` is a more fundamental operation rather than being defined in terms of `begin`, and `begin` can sometimes be defined in terms of it. This simplifies the definition of `array` and `vector`-like types at the cost of somewhat complicating `flat_map`-like types.